refactor: enhance login/logout tests by consolidating mock functions and improving clarity
